+++ openflowjava/openflow-protocol-impl/src/main/java/org/opendaylight/openflowjava/protocol/impl/core/VersionMessageUdpWrapper.java
@@ -5,11 +5,11 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowjava.protocol.impl.core;
 
 import io.netty.buffer.ByteBuf;
 import java.net.InetSocketAddress;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Wraps received messages (includes version) and sender address.
@@ -26,7 +26,7 @@ public class VersionMessageUdpWrapper extends VersionMessageWrapper {
      * @param messageBuffer ByteBuf containing binary message
      * @param address sender address
      */
-    public VersionMessageUdpWrapper(final short version, final ByteBuf messageBuffer, final InetSocketAddress address) {
+    public VersionMessageUdpWrapper(final Uint8 version, final ByteBuf messageBuffer, final InetSocketAddress address) {
         super(version, messageBuffer);
         this.address = address;
     }
