+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/services/util/RequestInputUtils.java
@@ -11,16 +11,16 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev13
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.MultipartType;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.MultipartRequestInputBuilder;
 import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public final class RequestInputUtils {
-
     private RequestInputUtils() {
         throw new IllegalStateException("This class should not be instantiated.");
     }
 
     public static MultipartRequestInputBuilder createMultipartHeader(final MultipartType multipart,
                                                                      final Uint32 xid,
-                                                                     final Short version) {
+                                                                     final Uint8 version) {
         MultipartRequestInputBuilder mprInput = new MultipartRequestInputBuilder();
         mprInput.setType(multipart);
         mprInput.setVersion(version);
@@ -28,5 +28,4 @@ public final class RequestInputUtils {
         mprInput.setFlags(new MultipartRequestFlags(false));
         return mprInput;
     }
-
 }
