+++ openflowplugin/src/main/java/org/opendaylight/openflowplugin/openflow/md/core/sal/convertor/data/XidConvertorData.java
@@ -5,10 +5,10 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.openflow.md.core.sal.convertor.data;
 
 import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Convertor data used in containing Openflow version and XID.
@@ -21,7 +21,7 @@ public class XidConvertorData extends VersionDatapathIdConvertorData {
      *
      * @param version the version
      */
-    public XidConvertorData(final short version) {
+    public XidConvertorData(final Uint8 version) {
         super(version);
     }
